pyproject.toml
[project]
name = "mcp-server"
version = "1.0.0"
description = "MCP Research Server with unified search and scrape"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "crawl4ai>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
    "readability-lxml>=0.8.1",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "trafilatura>=1.6.0",
    "seleniumbase>=4.0.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "hiredis>=2.3.0",
    "flower>=2.0.1",
    "fastmcp>=0.1.0",
    "mcp>=0.9.0",
    "pymupdf>=1.24.0",
]

[tool.uv]
dev-dependencies = []
